Rename DownloadedFile -> FileStorageFile
